Problem: moderator methods should be in same file
Solution: move moderator Currency methods to Currency file
