+++ lib/database/Currencies.js
@@ -1,12 +1,32 @@
 import { Mongo } from 'meteor/mongo';
 import { ActivityLog } from './ActivityLog.js';
+
 export const Currencies = new Mongo.Collection('currencies');
 export const PendingCurrencies = new Mongo.Collection('pendingcurrencies');
 export const RejectedCurrencies = new Mongo.Collection('rejectedcurrencies');
 
 if (Meteor.isServer) {
+  import { UserData } from '../../server/serverdb/UserData.js';
   Meteor.methods({
+    approveCurrency: function(currencyId) {
+      if(UserData.findOne({_id: Meteor.user()._id}).moderator) {
+          console.log("approvecurrencymethod")
+          var original = PendingCurrencies.findOne({_id: currencyId});
+          if (original.owner == Meteor.user()._id) {
+            throw new Meteor.Error("Approving your own currency is no fun!")
+          }
+          var insert = _.extend(original, {
+            approvedBy: Meteor.user()._id,
+            approvedTime: new Date().getTime()
+          })
+          Currencies.insert(insert, function(error, result) {
+            if (!error) {
+              PendingCurrencies.remove(currencyId)
+            }
+          });
+      }
 
+    },
     rejectCurrency(name, id, owner, message, moderator) {
       if(UserData.findOne({_id: Meteor.user()._id}).moderator) {
       var original = PendingCurrencies.findOne({_id: id});
@@ -33,7 +53,7 @@ if (Meteor.isServer) {
     }
   });
 
-  import { UserData } from '../../server/serverdb/UserData.js';
+  //import { UserData } from '../../server/serverdb/UserData.js';
   Meteor.publish('approvedcurrencies', function currenciesPublication() {
     return Currencies.find();
   });
